Recheck long running e2e tests are stable before merge
Since we rerun the github e2e tests (the fast ones) async, the time
between the check for stability of the slow google internal e2e tests
and when we are ready to merge could be quite large. Recheck the state
of the slow e2e tests before we merge.