Remove PRs from E2E queue if the fail on a later loop
It was possible that a PR would get on the queue, and while it was
waiting something would change. A subsequest loop of the munger would
change the status, but the PR would remain on the e2e queue. Remove it
from the queue as soon as we determine it shouldn't be there.