mungegithub: delete bot comments when they are no longer useful
Almost all of the bot's comments serve a purpose that can be resolved.
Once they are resolved the comments no longer need to be present and
just make review more difficult.

This PR removes the comments created by this bot. It does NOT remove
comments created by the cherrypick or the k8s-bot.

Resolves: https://github.com/kubernetes/kubernetes/issues/24007