Get new PR if interrupted PR has changed
If the interrupted PR has changed, it means we will need to re-test
it. Make sure that we don't test this PR if we have any other higher
priority PR to test first. This is done by going trough the loop on more
time without removing the interrupted PR from the list.