Do not update merge rate if we don't run e2e
If we don't run e2e tests, it makes it very easy to merge quickly, and
that makes the merge rate very high. Just do not include non-e2e merges
into the rate so that we keep it relevant.

This was the behavior before c548160bce (Create helper functions to
merge pull-request, 20016-05-23), where I incorrectly assumed this was a
bad copy-paste.