Remove labels before updating issue
Gorm is not smart enough to remove labels from the list of labels
associated to an Issue. As a consequence, if an issue is tagged with P0,
but is then unlabeled and relabeled with P2, we end-up with both tags.

As a workaround, we can remove all the labels from the DB before
updating an issue.