Remove broken 'use ${DB}' MySql statement
For some reason, the Golang MySQL driver doesn't work very well when
using 'use DB' statement (probably because it opens multiple concurrent
connections). Still we can't create the connection directly to the DB in
case it doesn't exist.

Create the connection, create the DB if it doesn't exist, disconnect and
then reconnect using the DB parameter.