Quit early if database fails
There are situations where we can't find when was the last item we
inserted in the database. If that happens, we restart fetching from
github from scratch (all the pages and everything). This obviously takes
a long time and is usually useless (if we can't read the last item
inserted, we may not be able to write anything).

Just quit early if we can't read the time of the last item. Obviously if
there is no item in the DB (first time), this is not an error and it
shouldn't fail.