Use ListFiles to calculate pull-request size
We currently use GetCommits to calculate the pull-request size. What
this does is that it requests the list of all commits in a pull-request
from Github, and then for each commit requests the list of all modified
files.

We can leverage Github "ListFiles" function that returns the list of
files with a single (beside pagination) request to Github. It's also
probably going to be faster.