Create auto-deployment for Google Cloud SQL
This patch makes it very easy to deploy a Google Cloud SQL along with
the cloud-sql-proxy container.

It creates a new github-database in your gcloud project if you don't
have already one (you can make it manually before to override the
default). It also creates a docker container with the required
credentials to connect.

All you have to do is to download the credential file, put it there,
create the conf with sensible values, and run deployx