Add GCI testing for kubernetes-e2e-gce-features tests
Job,GCE Project
`gci-gce-ingress`,`kubernetes-gci-ingress`
`gci-gce-flannel`,`kubernetes-gci-flannel`
`gci-gce-es-logging`,`kubernetes-gci-es-logging`
`gci-gce-petset`,`kubernetes-gci-petset`

NOTE: The flannel test is currently disabled. I cloned it, also kept the
clone disabled, and created the GCE project to run it when it is time to
enable it. Theoretically, you *should* just be able to enable it and
have things play nice when the time comes.