velodrome: Mix comments and events sorted by date
Currently, we receive all events and then all comments. It makes it
pretty impossible to measure things that can happen either as comment or
as events.

As an example, the first_comment plugins tries to find the first comment
or LGTM label. The problem is that it sees events first, and determines
that LGTM always happens first, even though it may happen last,
reporting incorrect data.

Fix that by merge-sorting both list when we get them from the database,
and push them to plugins in a sorted way.