velodrome: Set a memory limit to influxdb
InfluxDb is eating a lot of memories depending on what you ask it to
do. The problem is that if it starts eating all of the machine's memory,
random process starts to be evicted, including Kubernetes process. That
brings the machine down/off the cluster.

By setting a limit, the InfluxDB process will be evicted before it can
mess-up with the rest of the machine.