Adds Canonical Kubernetes e2e test results bucket
We've just started running the e2e gambit on a regular basis with
published results to the google storage bucket. This should enable
gubernator to ingest our test results to the federated dashboard

This is early work, there will be additional clouds incoming as we land
the CI jobs to do them.

 - Adds CDK to the testgrid config, + single CDK tab

- Tweaked the path per ixdy's feedback