Implement batched merging for the submit queue.
A commandline option specifying a Prow data endpoint is used to collect
results for batch tests. If a batch has passed all necessary tests and
is applicable given the current state of the repo, it will merge all
included PRs at once, without individually testing them.

This batch goroutine runs in parallel with the normal goroutine that
processes PRs sequentially. A mutex guards the critical section for merges
to avoid conflicts.