Relax OWNERSHIP rules if none found
It was possible that a PR changing files whose only "leaf" owner was the
PR author would result in 1 OWNER. The PR author. That OWNER would be
eliminated. Thus the PR would get no owner.

The fix is to look for ALL owners of files if we can't find any.