Add gci-gce testing against the GCI canary image
This should add testing against the latest GCI canary image. I am adding
one job first, to see if the way I did this works properly. The plan is
to add just a few more (gke and slow equivs) so we have some basic
signals from the very tip of GCI development.

Note that this depends on:
https://github.com/kubernetes/kubernetes/pull/38640