Fix docs and !GateApprovel
Fix the "docs" to show that approval either is or isn't
required. These are displayed on the info tab of the bot.

Even if `GateApproved` was false, if a PR had the approved label and had been
changed later we would still reject it.

This PR attempts to make the logic a bit more clear and the code a bit
more simple. Although we still do not have any tests for the
GateApproved=false case.