Test GCI tip against K8s tip on GKE
This should add testing against the latest GCI canary image. I am adding
one job first, to see if the way I did this works properly. The plan is
to add just a few more (rest of the gke equivs to the existing tip/tip
tests) so we have some basic signals on GKE from the very tip of
GCI development.