velodrome: Move prometheus argument to configuration
Right now, we use the default arguments from the docker container when
we run prometheus. The problem is that if we want to add new parameters,
then it will completely override this list. It also creates an implicit
coupling between the path used in the Dockerfile and the one in the
kubernetes configuration.

Use the arguments from the Dockerfile inside the kubernetes
configuration to remove the confusion.