mungegithub: Allow multiple commands in a comment
We currently do not allow multiple commands, or commands in the middle
of a comment. This change requires a small change in the API (we can
return a list of commands from a comment rather than just a pointer to a
command) but hopefully we don't have that many callers anyway.

This is mostly done by changing the regex to consider each new line as
the beginning of a new comment.