mungegithub: Replace AssignPR with AddAssignee
AssignPR has a very bad semantic: does it replace the current list of
assignees with only one person? Does it add a new assignee. Turns out it
does the former, and this unassigned a ton of people on all issues.

Let's make it more straightforward by using the non-deprecated function
AddAssignee