Disable CRI in CRI jobs
We have already enabled CRI in all builds. This change repurposes the
CRI jobs to test the original non-CRI implementation. We keep the
original gce project (and job names) to avoid recreating all the
projects. This could be misleading at times, so this change renames
all "cri" jobs to "non-cri" to improve readability.