Explicitly specify host cluster.
We still track the names of the clusters through their GCE zone names
in the core repository's test scripts. So we are using the cluster
zone as a proxy to identify the cluster. In the future we are going
to have a way to directly identify clusters through their names.