kops: Move kops-aws-updown job to its own isolated account
This is a re-drive of #1843 after kops was shifted to a scenario/env.

Still TODO: The wheel of availability zones in
scenarios/kubernetes_kops_aws.py needs to be moved somewhere else -
each account doesn't have access to the same AZs. However, if we just
keep the primary account for presubmits and separate the postsubmit
tests, we really don't need this functionality. Locked kops-updown to
a us-west-2 AZ that happens to be available to the account.