Default kubernetes-anywhere to use the "latest" Kubernetes version.
Previously, this phase2.kubernetes_version field was only referenced if
the phase2 provider was "ignition," and completely ignored if the phase2
provider was "kubeadm." Now, upstream changes in kubernetes-anywhere
make the kubeadm provider use this value to determine what Kubernetes
images to use.

Using the string "latest" matches the previous behavior, since this is
the kubeadm default if no Kubernetes version is specified.