approvers: Track "Approval" rather than "Approver"
Currently, the code keeps track of who has approved a pull-request. We
know want to track slightly more than that, including: how did the
person approve the pull-request (might be through lgtm, might be
approval, might be because they're the author), and also when they did.

This forces a somewhat important refactoring, also enforcing better
encapsulation of some structures.