approvers: Consider assignees when CCing
Right now we don't really make a corelation between the people assigned
to a PR and the people we select for approval. This is a problem because
some time we could reduce the list of suggested people by looking at the
list of assignees.

Consider the list of assignees when we suggest people, and don't suggest
people that can't approve more than current assignees.