velodrome: transform: Remove series on restart
We use to find the last point and then not re-insert data before that
last point. This is actually a burden to maintain when one wants to
change the way metrics are computed, because the transform plugin needs
to be stopped, then the measurement needs to be dropped, and then
transform needs to be restarted. These steps adds-up to too much
maintenance.

Let's just drop the series automatically on restart. This shouldn't have
too much impact on the quality because:
- Restarts don't happen so often
- Even though the loop may take some time to run, it's very small
compared to the lifespan of transform.