velodrome: Store IDs as strings
The problem is that Gorm is trying to be smart and creates
auto-increment primary key when you have ID as an int. And the problem
is that it won't let you create a composite index in that
situation. Make every ID a string so that we can make them part of a
composite primary key.