Disable parallel test runs for kubeadm e2e job.
Actual test execution is still currently disabled, but I noticed that
this setting was creating flakes in local testing because some of the
Conformance tests are also marked Serial and can't safely be run in
parallel.

Instead of skipping the Serial tests, let's just disable parallelization
for now to be more thorough in the testing.