velodrome: fetcher: Download comments separately
There used to be a bug in github that prevented us from downloading
comments directly. The workaround was that we were downloading comments
for each individual issue rather than all comments at once.

This has been fixed on Github so we can now use that API again. It means
that we don't need to make 2 github calls per issue we update, which
makes it much faster and much better on token usage.