velodrome: transform: Filter comments/events by date
We currently only fetch comments/events that have been updated after a
specific id, the problem is that IDs are not sorted properly because
they are strings (rather than int). Let's filter by date time and add a
new index to make it fast (as we sort by date, also because ids are not ints).