approval: Search all possible approvals for assignees
Currently, even if an assignee can approve everything (because for
example he's a root approver), we can't use him as a suggested person
because suggestion only looks for leafApprovers. When we consider who is
the best assignee to suggest, we should not only look for leaves, but
for all possible approvals.

One of the consequence of that is that if a root approver is assigned,
he'll almost always be suggested.