velodrome: fetcher: Fetch events for each issues
Github doesn't let us download all events for the repository anymore. As
a work-around, we can download events for each individual issue. This is
not as great because it means that for `n` issues, we need to make
`O(n)` calls to github API rather than `O(1)`, but all of this is done
completely in the background so it doesn't really matter after all.