Quick fix for kubeadm e2es.
Until the sustainable fix is in place, I have a short-term workaround in
my fork of kubernetes-anywhere. I'd like to switch to using that in
order to get the kubeadm e2es stable, then figure out the correct fix
for kubernetes-anywhere and revert this commit.

This also re-enables Conformance testing.

https://github.com/kubernetes/kubeadm/issues/219