Tidy up our Bazel rules.
Add requests and yaml repositories. After this change, the test_history
page and bootstrap_test should have no external dependencies. We could
almost get rid of the requirements file for test_history, but it's
actually depended on in a few other places, such as kettle.

Make //kettle:make_json_test pass on my laptop by changing a "python" to
a "python2". This call is problematic, since the subprocess is executed
outside of Bazel's carefully crafted sandbox, and won't see the yaml
package from the rules.