Allow disabling automatic --repo flag in bazel runner.
If a periodic job uses the bazelbuild image and passes an argument of
"--repo=...", it ends up getting appended to the list of repos instead
of overwriting it. This change lets jobs set an environment variable to
opt out of this automatic --repo flag added by the bazel runner.

This also reverts the attempted (failed) workaround from
https://github.com/kubernetes/test-infra/pull/2393.