Fix e2e-kubeadm release-1.6 job.
Recent changes to kubernetes/master moved where bazel .deb files get
pushed. The kubernetes_e2e.py scenario was updated to accommodate this
change, but the kubernetes/release-1.6 branch still pushes these .debs
to the old location. Since we only support ci/periodic jobs for
release-1.6, and we expect to use the new location from now on, consider
this a special case for now and we can remove this code when we no
longer support the 1.6.x series.