Make the federation presubmit job blocking.
It has been a little over a week since we started reporting the results
of this job on all the PRs. We did not have any major issues. We had a
couple of minor hiccups: Issues https://github.com/kubernetes/kubernetes/issues/45795
and https://github.com/kubernetes/kubernetes/issues/45978. We had foreseen
problems of the first type but the second one was a little surprising.

SIG-Federation is starting a buildcop rotation and the buildcops should
be able to handle both these types of situations. We don't have all the
tooling in place for non-Googlers to handle these issues because it
needs access to Jenkins, so they still need to ping a Googler when
they see a problem. We are working on moving these jobs out of Jenkins
to prow (Issue #2791).

Empirically, these problems have been uncommon and shouldn't affect
the submit queue often.