only-test-affected: Run bazel query once for all files
The current implementation is running bazel query once for each file
that has changed. If the repository is big, that can sum-up to a lot of
time. Change that O(n) to O(1) by calling bazel query on the entire set
directly.

As an example, each call to bazel query takes about one second with my
machine. A 60 file diff takes a minute to run before that patch, and 1
second after it.