Updated submit queue Makefile and improved templatization.
Changes to the Makefile make it easier to deploy a submit queue to a new
cluster.
Changes to templatization fix the problem with not all targets requiring
the same secrets. I also templatized secrets, volumes, volume claims,
and services.  The templatization changes should make it easy to switch
to using helm charts.
Finally, I moved some of the directories out of mungegithub/submit-queue/deployments as they were not standard submit-queue deployments.