Fixed cherrypick log search error.
The cherrypick app determines if a cherrypick is complete for a PR by
checking if the git log for the appropriate release branch contains the
changes made by the PR. `git log --grep` is used to check if the logs
contain the commit messages. This form of git log expects a regex
pattern but we provide a string literal pattern which causes an error
when the literal is not a valid regex. Adding the -F flag matches
literals instead of patterns.