Updated cherrypick log matching for script generated picks.
Currently the cherrypick app identifies cherrypicks merged into release
branches with hack/cherry_pick_pull.sh via the text 'Cherry pick of <pr
num> on release-<release version>.' This text is used in the merge
commit messages made by the submit-queue when it makes automatic merges
since the text is in the PR body which the submit-queue uses to craft
the message. However, if a human merges the cherrypick PR into the
release branch, only the title of the PR is used by default in the
merge commit message so the message will not contain the text we are
looking for and the cherrypicker wil not identify that the PR was merged.
By searching for the text used as the issue title we will always find
the merge commit unless a human manually merges and modifies the merge
commit message.