Add a git client to prow.
This will let us write plugins that would like a local git clone to mess
around with. It uses a cache so that multiple clones of the same repo
will be quick.

Note that this makes the hook image dependent on the git tool. This is a
shame, but it's unavoidable.

Right now the client only knows how to clone repos. In future work we
can add methods to the Repo struct that will let us do more interesting
things.