Correct the --extract arg for gci-ci-* job configs
These jobs test a daily build of GCI with a particular branch head of
k8s. According extract.go, the correct setting should be
`--extract=gci/<gci-image-family>/<k8s-ci-version>`.

Also, remove uses of the  env var JENKINS_GCI_HEAD_IMAGE_FAMILY that is
now obsolete. The only remaining use of it is in
`jobs/ci-kubernetes-e2e-gke-gci-ci-master.env`. I am not sure what that
job is used for so I'm not touching it.