kubetest gke: Move DumpClusterLogs to the deployer, implement
This adds a DumpClusterLogs in the deployer interface in a
kinda-polymorphic way - most of the interfaces just thunk to a
defaultDumpClusterLogs that does what we did before, and I implemented
a new, extra hacky one for GKE.

The hacky one for GKE uses the custom function interface I added for
kops. This whole area is a big TODO, though - we need to get this out
of k8s/cluster/log-dump.sh.