Bring hook plugin configuration into plugins.yaml
As more and more plugins gain configuration options, it is becoming
clear that the overall Prow configuration is not the appropriate place
for these options. Configuration options for these plugins are almost
always created in order to support more organizations or projects, so
the mapping of project --> plugins we already have in config.yaml is
the correct place to put these options. With this patch, that config
file now maps project --> plugins and config. Config for the repository
being operated on is exposed to handlers at run-time.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>