Refactor matchers to take the munger's name as config
For the same justification as in 022f93debc369476aada221e07ee7e952fa9c38d,
it is necessary to be able to configure the name of the robot that is
munging GitHub data. This patch threads through this information into
the various places the "k8s-merge-robot" was hard-coded in the matchers.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>