Split primary kubernetes/kubernetes submit-queue munger into its own instance.
This will allow faster mungegithub development with less downtime of
the primary service (the submit queue). This makes the submit queue
asynchronous to mungers running in the other queue (like removing lgtm
after a commit), but in practice this shouldn't be a problem as long as
the presubmit time is longer than the other queue's loop time.