Subsume the aliases feature into repo-updates
We want to be able to providea stricter and more useful API on the
repo-updates feature. Namely, we would like to provide lists of
approvers and reviewers that are case-insensitive and have their aliases
resolved from the outset, as the non-resolved and case sensitive data is
not terribly useful to the end user in a munger. In order to provide
this data without linking two features together, we need to subsume the
alias feature into the repo-updates feature so that these processing
steps may happen under the hood and invisibly to the consumer.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>