Allow for author self-approval to be toggled with config
When a PR is working on a section of code that is not covered by the
default set of blocking jobs, and the author is able to approve their
own PR, today anyone in the organization can drive-by with a `/lgtm` or
otherwise innocently place an `/lgtm` that causes the PR to merge even
if it is not ready, and the only method the author has to stop this is
to add and remove the `do-not-merge` label. Whether or not an author
immediately and implicitly self-approves their PR should be configurable
to allow for organizations to turn this feature off and improve the UX
of the approval manager for developers in these cases.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>