mungegithub/publisher: scripts: process merged PR by PR and merge on first merge commit on PR branch
The algorithm is switched from a pure git-log loop to a loop over the commits on the main
line (--first-parent) of the filter-branch'ed repo.

For each merge commit on the main line, another local loop is started over the PR branch
commits. If a merge commit is found there, only the commits above it are cherry-pick
one-by-one. Everything before the merge commit is squashed and commited. This allows us
to survive merge commits while loosing commit history. The later is better than failing
completely. In the future we want to forbid merges in PRs and therefore avoid missed history
during publishing.

This PR also removes the kubernetes-sha file, publishes the Github merge commits. Moreover,
tags are added to Github merge commits.