Expand the do-not-merge label into a class of labels
In order to support a larger set of reasons why a developer or the
system may wish to stop a pull request from merging and in order to do
so in a way that is clearly communicated, the `do-not-merge` label is
broken into a `do-not-merge/X` label class. The submit-queue will not
merge a PR if any label with the prefix exists. This change is
backwards-compatible as the original `do-not-merge` label matches the
prefix and no munger was removing the `do-not-merge` labels they added.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>