Consider comments using deprecated labels stale as well
After bc2bc3c the comments that these mungers add have changed, but when
we encounter a comment with the previous format, we should consider it
stale under the same conditions that we would consider a comment with
the new format stale. This will ensure that all of the comments left
before the change went live get cleaned up as necessary.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>