Make the Prow bump script smarter
The script will now automatically detect which components need to be
bumped and determine the last time they were bumped to ensure that it is
not easy to forget to bump. Furthermore, this allows for a daily bump to
occur that will pick up all the changes that ocurred during that day.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>