Only mutate the required contexts when changing branch protections
The configuration for the mungers today only allows for the required
contexts to be explicitly stated for a protected branch. Until we expose
the other types of protections for a branch, our branch protection
requests should make sure not to mutate those fields. Furthermore, now
that we know there are types of protections we do not control, it is not
possible to know that we should remove a branch protection entirely, so
that functionality is removed. If all types of protections are exposed
through the configuration in the future, that functionality can return.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>