Add external PR status refresh plugin
This is an external prow plugin that helps updating stale Github
contexts in PRs. We had these so far because plank and the
Jenkins operator were not respecting termination signals but that
was a separate issue that got fixed and stale contexts are still
a possibility in case Github breaks down.

Signed-off-by: Michalis Kargakis <mkargaki@redhat.com>