Log dump: support for dumping additional IPs
DumpAllNodes now takes an additional list of IP addresses.  Dumping
through kubectl get nodes is still preferred, but if an IP address is
specified that was not found in that list of nodes, then that IP address
will be dumped also.

This allows for the installation tool to provide an additional list of
IPs that is finds e.g. via cloud APIs, and that thus log dumping will
still work for nodes that fail to register with the cluster (or even if
the entire cluster fails to come up)