Expose the ProwJob spec through environment
When building tools that run in the test pod or in test jobs, it is
critical to be able to determine the ProwJob specification that was used
to trigger the job in order to determine where the appropriate GCS
bucket is, which repositories need to be checked out, etc. In addition
to providing this data in a disparate set of environment variables, we
should also just provide the ProwJob spec in serialized form so that Go
programs importing test-infra can simply unmarshal it and go from there.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>