mungegithub/publisher: merge commit list to pick without branch commits
The cherry-picking algorithm depends on the following assumption:

  If the PR head merges cleanly into master, all the PR branch commits
  cherry-pick cleanly onto the PR that merged just before.

Unfortunately, this is not always true. Here we add a list of PR merge
commit hashes where we do not try to cherry-pick the PR branch commits
at all, but only the first parent of the merge commit. This gives the
same result, but we loose history. But better we loose history, than not
being able to publish anymore.