Switch kops newrunner jobs to use alpha channel
The alpha channel contains the 'development' configuration - in
particular the AMIs / Google Cloud Image to use for each k8s version.

We want to be using the alpha channel for tests that aren't tied to a
release branch - both so that we test the upcoming images, but also
because the images for upcoming kubernetes versions won't necessarily be
in the stable channel.