Fix handling of GOPATH which is now optional since Go 1.8
Much of this code depended on $GOPATH being set and would fail with hard
to debug messages otherwise.

Since Go 1.8 the $GOPATH environment variable is no longer necessary and
will default to ~/go (on Unix and Linux systems.)

Update this code to use the "go/build" package and more specifically the
build.Default context to find the GOPATH. (Tested this with a $GOPATH
set and confirmed it reflects it.)

Furthermore, split paths (in case $GOPATH contains multiple paths
separated by a ":" delimiter) and use the appropriate one in case there
are more than one. In that case, look for ${gopath}/src/k8s.io/${topdir}
and pick the first one that exists. If none of them do, default to using
the first entry in $GOPATH.

Tested: by running `make test-e2e` from the top of Kubernetes repository
with $GOPATH unset in the environment. Confirmed that the
`make -C ${k8spath} quick-release` command spawn by kubetest would use
an absolute path to my ~/go tree.