Disables SELinux for Planter container
As Planter uses a bind mount on the /home/lmadsen directory, on systems that use
SELinux, the bind mount requires SELinux to be disabled so that files can
be accessed. We're unable to use the :z or :Z options, as this causes a
relabling of the entire /home/lmadsen directory contents, which has the implication
of locking admins from sshing back into the machine.

By creating the Planter container as a Super Privileged Container we
avoid administrators from having to disable SELinux system wide.

Relates to #5607