Adding capability for HA control plane testing with kops
Added a new flag for kops-ha. Also, since we now may run in regions that
may not have m3 instances, updated the master type to c4 instances.

In order to test HA control plain a minimum of 3 AZs must exist.
Implemented a func that queries all regions, and then queries the list
of that regions AZ. A random region is then selected. If running in ha
mode, a region with a minimum of 3 AZ is selected. The new zone
capability can be run in both ha and non-ha modes. If --kops-zones is
not set, and the test run is on AWS, the new code is used. You can
override the new code by setting the zone as done previously.

Added a unit test, that requires a live AWS account for testing the new
random region selection. These tests are skipped since they require a
live AWS account.