Implement a tool for reporting job start status
The `initupload` tool uses the Prow-provided environment in `$JOB_SPEC`
as well as clone records from disk to determine how the initalization of
the job has occurred and where metadata about the job should be uploaded
to reflect that. This tool will always upload `started.json` to the
appropriate location; if the clone records show that a step has failed,
the tool will also push `finished.json` as well as `build-log.txt`,
formatted to contain the clone records in a human-readable format.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>