Fix check for the SSH key for the project being set up.
This prevents running `config-ssh` when it is not actually needed.

This was actually intended, however the check had a bug in that the
order of arguments to strings.Contain() was incorrect. :-)

Tested: manually by running kubetest --up with the appropriate arguments
and checking that my ~/.ssh/config wasn't touched afterwards.