Introduce the entrypoint and sidecar containers
In order to wrap the test execution, two new utilies are necessary. The
entrypoint container executes the test process and redirects the output
from `std{out,err}` to a file. When the test process has finished, the
entrypoint writes the test's exit code to a marker file. The sidecar
container waits and watches for the marker file to be present and, when
it is, generates `finished.json` from the exit code and uploads that,
the process log and any artifacts to GCS using the `$JOB_SPEC`.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>