Refactor pod utils to take GCS args using JSON
When the pod utilities are being launched by a program like Prow, it
does not make sense to take rich, structured data and flatten it into
simple strings to pass arguments between the two programs. We can use a
JSON configuration file to pass the minimal set of configuration options
necessary for the tool to function. This option can coexist with the
current set of flags, which will be useful if a human is to use the tool.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>