Introduce a tool to create a Pod from a ProwJob
When testing the behavior of `plank`, it is very useful to be able to
run only the logic that transforms a `ProwJob` to a `Pod`. The `mkpod`
tool allows for this; in concert with the `mkpj` tool, we now are able
to create a `Pod` for a job in the Prow configuration without running
any part of the larger system.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>