Use 1.8 and 1.9 branch heads for GCI testing and retire old jobs
This a variant of https://github.com/kubernetes/test-infra/issues/7646.
The GCI test jobs always pin to the built-in version of k8s, which are
broken by the newer gcloud cli used by the infra to launch jobs. Both
1.8 and 1.9 have been patched, but no release has been made so far. This
change uses the `gciCi` extract strategy to pick up the latest CI
version on the older branches so the tests will pass again.

Also, remove jobs for GCI m63 which is deprecated now.