Add ProwJob name as a label on Pods
Complex workloads on test/build clusters where more than one Kubernetes
object is used to fulfill a test request should be possible for
downstream consumers of Prow. Instead of enforcing that the object names
must match the `ProwJob` ID, we should instead allow for a label to be
used to identify these objects as well as the original method.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>