Remove most gci image test jobs
They have all been migrated to another Prow instance, but before we get
a replacement for the BQ table with test results, it's better to keep at
least the release blockers (gci-qa-m*) for a little longer.

Also, take more projects out of the `gci-qa-project` pool now that we
have only four jobs here.