Hide error when checking for network/firewall
Up() and ensureFirewall() use control.FinishRunning() with gcloud ... describe to check whether a resource exists or not. However, since FinishRunning displays the output of the command, if the resource doesn't exist the log will display a misleading "ERROR" entry.

Use control.NoOutput() instead of FinishRunning() to hide the output (and thus the misleading error message), but still check if the resource exists or not based on the return value of the command.

control.NoOutput() is like control.Output(), but it completely discards the output of the command.