kops e2e: stop hard-coding c4.large
Otherwise we're not testing the kops behaviour: kops can auto-detect
whether instance types should be available in certain regions, and
chooses between m3 and c4 instances for the master.

Note that this isn't (currently) stock-out detection, but stock-outs are
a problem in kops as well so we want to get them under testing so we can
fix them.