prow: add stage plugin
The stage plugin applies the stage/alpha, stage/beta,
stage/stable labels to issues in the kubernetes/features repo.

Since all the stage labels are mutually exclusive, only one stage
label is allowed to persist at a time. If one stage label is applied,
other stage labels are removed (similar to the lifeycle label).

Note that this commit does _not_ enable the plugin. If we enable the
plugin in the same PR it is created, there are chances of the plugin
not getting enabled (because the plugin will not exist in the prow
that is currently deployed).