Make Tide omit PRs failing GitHub contexts from the pool.
Previously PRs that were failing, but otherwise mergeable were included
in the pool. Now they are filtered out of the pool before subpools are
synced and before the pool is given to the status controller.
Importantly, we allow PRs with unsuccessful statuses to remain in the
pool so long as all unsuccessful statuses are pending contexts
associated with ProwJobs that are required for merge.