prow: trigger: skip jobs whose parents are skipped
We must provide a status for every ProwJob that is configured and not
marked to be silent, both so that developers can see what happened to
their PR as well as so that tide can determine what the status of the PR
is. Therefore, we have to mark contexts that are children of skipped
contexts as also skipped.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>