Avoid highlighting pathological logs.
Extremely long logs with many, many mismatches are going to result
in the creation of a ridiculous number of DOM nodes. Browsers generally
do not respond well to hundreds of thousands of DOM nodes suddenly
appearing, and so hang.

Limit the number of mismatch spans to 1,000 (several times the observed
number on any reasonable test case) before we just show the
unhighlighted log.