Add conditional enabling of IPv6 for bootstrap container
This change adds a mechanism for enabling IPv6 in any Prow containers that
are built from the bootstrap container, based on an "IPV6_ENABLED"
environment variable that can be passed to the container.

This mechanism will be used by an IPv6 CI test that will be based on
kubekins-e2e (see PR #7529).