kubeadm: pull kubernetes-anywhere per job
kubernetes-anywhere (k-a) now has a specific branch for use with
e2e tests and it is no longer needed to bake a k-a commit SHA into
the kubeadm image.

The following is used to pull the `kubeadm-e2e` branch of k-a:
`--repo=k8s.io/kubernetes-anywhere=kubeadm-e2e`

This removes a step where after fixing a k-a bug, the test-infra
kubeadm image has to be updated.

Also removes the kubeadm runner as it is no longer needed.