Disable the cache mutation detector.
`ENABLE_CACHE_MUTATION_DETECTOR` has been set for over a year, but due
to a bug in the script parsing it, has not actually run. Fixing the bug
such that it can run makes the tests time out. Since an hour is already
a long time for presubmit tests, let's not make them longer and instead
just stop enabling this.

See kubernetes/kubernetes#67919 for more context.