Prow: Improve labelling resources with job names
Previously, `pjutil` attempted to label ProwJob resources with job
name, but when the job name was longer than 63 characters, the label was
not applied (because labels have a 63 character cap).

Apply truncated job name as a label in this case, which is better than
no label. Full job name is now always present as an annotation with the
same key, just like `pod-utils/decorate` does this.