tide: Add merge_github_approve_tag config setting
Make it easier to find out who approved a given commit/merge by
appending that information to the commit message.  Homu has something
similar to append 'Approved by' [1] when it merges a pull request.
And Git [2] and Linux [3] both define a Reviewed-by tag with similar
semantics to GitHub approval.

This is pulling approvers from GitHub because we can do that with
straightforward GraphQL.  I've used "GitHub" (or "github") in the YAML
and Go property names to make a clear distinction between this this
GitHub-review-based approver list and an approver list calculated via
the lgtm and/or approve plugins.  Tide itself does not care about
either GitHub reviews or the lgtm/approve plugins; Tide only cares
about pull-request labels.  By using GitHub reviews, we can lean on
GitHub to track the to-be-tagged approver set for us.  We may try to
couple the plugins later (e.g. [4]), but for now, I'll use GitHub
approvers and distinct names.

Using GitHub in the YAML and Go property names makes the distinction
between this

I've added unit tests for the rendering code.  Testing the GraphQL
structure itself seemed hard to do except with a possibly unreliable
mock.  So I wrote a one-off script for testing the GraphQL pull:

  package main

  import (
    "context"
    "fmt"
    "log"
    "os"

    githubql "github.com/shurcooL/githubv4"
    "golang.org/x/oauth2"
    "k8s.io/test-infra/prow/tide"
  )

  func main() {
    ctx := context.Background()
    src := oauth2.StaticTokenSource(
      &oauth2.Token{AccessToken: os.Getenv("GITHUB_TOKEN")},
    )
    httpClient := oauth2.NewClient(ctx, src)
    client := githubql.NewClient(httpClient)

    var query struct {
      Repository struct {
        PullRequest tide.PullRequest `graphql:"pullRequest(number: 46669)"`
      } `graphql:"repository(owner: \"kubernetes\", name: \"kubernetes\")"`
    }
    err := client.Query(ctx, &query, nil)
    if err != nil {
      log.Fatal(err)
    }

    fmt.Print(tide.ApproverTags(&query.Repository.PullRequest, "Reviewed-by"))
  }

To get that script running, I temporarily patched my function to be
the public ApproverTags instead of the private approverTags.  Then:

  $ GITHUB_TOKEN=... go run main.go
  Reviewed-by: Janet Kuo <janetkuo@users.noreply.github.com>
  Reviewed-by: krmayankk <krmayankk@users.noreply.github.com>

My initial intention was to inline the structure definitions, but I
ended up creating named types for most of them to make it easier to
write the compound literal for the test.  The new types are added in
roughly alphabetical order after the existing PullRequest type.

I have not used my new User type for PullRequest.Author.  The existing
code was fine getting only the Login there, so I'd only waste some
bandwidth transferring names and email addresses as well if I used
User there.

The updated CommitMessage comment is from GitHub's docs for the
associated property [5].  The previous wording sounded (at least to
me) like our string was the entire commit message, when in fact it's
just a footer on GitHub's automatic message.

The Author formatting avoids @ so we don't pester approvers if their
commits are cherry-picked, forked, etc. [6].  Otherwise, the
formatting follows Christoph's suggestion [7] to use the GitHub login
only as a fallback for missing name/email information.  That email
fallback assumes we're using github.com repositories (and not some
internal enterprise GitHub version), but that seems like a fairly safe
assumption for repos using Tide.

I'd personally prefer (*PullRequest).authorTags() and
(*Author).String() methods to the functions in this commit, but Cole
explicitly requested functions [8].

[1]: https://github.com/servo/homu/blob/461e9acbef5e984d18e0342b679ee424a945cfaf/homu/main.py#L722
[2]: https://github.com/git/git/blob/v2.19.0/Documentation/SubmittingPatches#L358-L361
[3]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/Documentation/process/submitting-patches.rst?h=v4.18#n565
[4]: https://github.com/kubernetes/test-infra/pull/9400#discussion_r218256429
[5]: https://developer.github.com/v3/pulls/#merge-a-pull-request-merge-button
[6]: https://github.com/kubernetes/test-infra/issues/8242#issuecomment-394528579
[7]: https://github.com/kubernetes/test-infra/pull/9400#discussion_r218209546
[8]: https://github.com/kubernetes/test-infra/pull/9400#discussion_r218253580