Ignore push events for deleted branches
When a branch is deleted, GitHub sends both a `delete` event
as well as a `push` event updating the branch to an invalid
commit SHA. No consumer of push events is expecting an invalid
commit at the HEAD; we should allow plugins to subscribe to
the `delete` event if they want to handle that case. So, if
we can detect that this is the weird `push` event before the
imminent branch deletion, we don't handle it.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>