Set state fields from PR when handling a review
Branch, body, and author fields were previously unset or incorrectly
sourced from the review instead of the PR. This removed the ability for
authors to implicitly self-approve because their authorship was replaced
by the author of the currently handled review.

One test for each of handleGenericComment, handleReview, and
handlePullRequest now verifies that the state given to handleFunc is
correct.