kubetest/eks: do not delete binaries on "Down" call
My assumption was "Down" is only called once at the very end of the
tests, but it turns out kubetest may call "Down" multiple times
while we download the binaries only once when we create EKS deployer
objects.

The downloaded artifacts will be gone anyway on pod termination.
This PR removes the "os.RemoveAll" calls, so that the downloaded
binaries can be reused.

Signed-off-by: Gyuho Lee <leegyuho@amazon.com>