OAuth: Invalidate GH session on failures instead of server errors
Previously, `HandlePrStatus` failed with 500 when the user has an
invalid/stale cookie. After the fix, the behavior is the same like when
the token is successfuly extracted from the cookie but fails to
authenticate with GH. The sessions are invalidated by deleting the
cookie and `HandlePrStatus` returns indicating the user is not properly
logged in with GitHub via OAuth.

Fixes: #10063