Ensure correct blocking presubmit state on config reload
A new component, the `status-reconciler`, is created that subscribes to
changes in the Prow configuration in order to ensure that changes to the
configuration do not cause pull requests in flight to get stuck in the
Tide queues. In order to do this, the controller determines if any
changes to the configuration edit the set of blocking presubmits for any
repo. There are three types of change, for each of which the
`status-reconciler` will make the appropriate action:

 - a new blocking presubmit context is added; a ProwJob is created for
   every trusted pull request in flight to populate the new status
 - a current blocking presubmit is removed; the GitHub statuses on any
   pull requests in flight are updated to mark the status as retired
 - a current blocking presubmit is migrated to a new status; the GitHub
   statuses on any pull requests in flight are updated to mark the new
   status that was migrated to

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>