sig-storage: initial config for kubernetes-csi/csi-lib-utils
kubernetes-csi/csi-lib-utils is a brand-new stand-alone repo with some
Go utility package (one at the moment, more to come). "make test"
inside it enforces coding style and that unit tests pass, which is
something that we want to ensure before merging any PR.

This is the first config for kubernetes-csi, hence the new OWNERS file.

Fixes: kubernetes-csi/csi-lib-utils#2