deck: perform initial tide sync asynchronously.
If tide is unavailable at deck startup (as it tends to be during a prow
upgrade), deck gets stuck waiting for it for potentially several
minutes, which is not a desirable user experience, and can be conerning
for the person actually performing the upgrade.

There is no particular reason that the tide status must be known at
startup (and if they fail it blindly starts up anyway), so move the
initial update to happen asynchronously.