Use Update() instead of UpdateStatus()
The prow.k8s.io deployment will not be on Kubernetes 1.13+ for a long
time, so we cannot declare a /status subresource on our CRDs and use it
until then. That approach is best practice but it will have to wait.
Using the Update() call is what the controllers used to do.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>