Expose errors in removing cached cronjob entries
The original use of the defer statement in a loop here was to get
around the concurrent use of the jobs map, but it did not bubble up the
error returned from the removal function. By computing the full list of
jobs to remove from the cache first we can remove the jobs without using
defer and ensure we bubble up the errors.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>