Trigger correct set of jobs on PR events
When a PR is opened or updated, the set of jobs that would automatically
run should be triggered. The previous implementation attempted to reuse
the comment-command triggering code to determine which jobs to trigger
by passing `/test all` as a synthetic comment. This was incorrect as the
comment command `/test all` not only triggers jobs automatically but
will also trigger jobs that have a trigger regex that matches the
command. This resulted in jobs that had a configuration like the
following being triggered on PR creation:

    always_run: false
    run_if_changed: nil
    trigger: ((?m)^/test( all| job-name),?(\s+|$))

These sorts of jobs should be triggered only manually with a comment
command from a user.

This change refactors the filtering mechanisms in order to allow us to
explicitly request the correct behavior instead of trying to achieve it
by reusing a comment command parsing mechanism.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>