Refactor cluster config loading, add unit tests
This code did not previously have unit tests but was a cornerstone of
using Prow with multiple build clusters, so this refactor shapes it in a
way that is testable. It is not possible to test the exported
`LoadClusterConfigs` function as it is not possible to have the
in-cluster configuration loader helper not pick up state from the test
environment. In lieu of testing that, every sub-component of the loading
chain is unit-tested as best as possible.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>