Catch invalid labels earlier
Previously the code to verify labels was only being run if
the --orgs flag was passed.  This allowed people to check in
invalid labels (description too long) and not find out about
it until the first time the label-sync job ran.

This change will validate all labels regardless of whether
--orgs is passed in, allowing the hack/update-labels.sh
script to catch invalid labels. The warning about repos not
being in --orgs is only triggered if --orgs is passed in
the first place