boskos: AWS janitor clean -all option
This patch updates the AWS janitor command with the new flag -all.
Executing the command with -all cleans all visible resources, bypassing
the mark and sweep logic. When -all is specified, the S3 -path flag is
ignored.

Additionally, the CleanAll function has been placed into the "resources"
package, and the function is now used by both the AWS janitor (when
-all) and the AWS janitor for Boskos.